Break out commands into own file
